+++ cmd/lerd/main.go
@@ -29,6 +29,7 @@ func main() {
 	root.AddCommand(cli.NewInstallCmd())
 	root.AddCommand(cli.NewStartCmd())
 	root.AddCommand(cli.NewStopCmd())
+	root.AddCommand(cli.NewQuitCmd())
 	root.AddCommand(cli.NewUpdateCmd(version.Version))
 	root.AddCommand(cli.NewUninstallCmd())
 	root.AddCommand(cli.NewParkCmd())